+++ Natasha/docs/release-notes.md
@@ -16,6 +16,7 @@
 * Completely changed the Access Violation card.
 * Changed the Reprogram card.
 * The Access Violation card is now a rare (was uncommon).
+* Changed visuals of the Hijack Engines card.
 * Changed visuals of the Voltage Tuning card.
 
 ## 1.2.0